+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.2.2
 - CSpect: timeout error if no response received.
+- Fixed: "TypeError: cannot read property 'getObject' of undefined"
 
 # 1.2.1
 - Renamed "Memory Pages" to "Slots".