+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# 1.2.3
+
+
 # 1.2.2
 - CSpect: timeout error if no response received.
 - Fixed: "TypeError: cannot read property 'getObject' of undefined"